Can unpload data from api to local db
