--- app.js
@@ -1,18 +1,0 @@
-// const express = require('express');
-
-// app = express();
-
-// const bodyParser = require('body-parser');
-
-// app.use(function(req, res, next) {
-//     res.header("Access-Control-Allow-Origin", "*");
-//     res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
-//     next();
-//   });
-// app.use(bodyParser.urlencoded({ extended: false }));
-// app.use(bodyParser.json());
-
-// console.log("lz");
-// app.use('/', require('./index'))
-
-// app.listen(3000, () => `listening on 3000`)
